requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据验证
pydantic>=2.10.0
pydantic-settings>=2.1.0

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver

# 缓存与消息队列
redis==5.0.1
aioredis==2.0.1

# HTTP 客户端
requests==2.31.0
httpx==0.26.0

# 音频处理
pydub==0.25.1
librosa==0.10.1  # 音频分析

# 存储
tos==2.6.0

# LLM
google-genai>=1.0.0  # 新版 SDK，支持结构化输出

# Azure Speech SDK
azure-cognitiveservices-speech==1.34.0

# 日志
structlog==24.1.0

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 工具
python-multipart==0.0.6  # 文件上传
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4  # 密码哈希

# 监控
prometheus-client==0.19.0
psutil==5.9.6  # 系统资源监控

# 测试
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.0  # 属性测试

# 代码质量
black==23.12.0
ruff==0.1.0
mypy==1.7.0
